alura/amazon_bedrock_python_construa chatbots_ferramentas_geracao_imagens_com_ia/requirements.txt
openai
python-dotenv
pydub
instabot
Pillow
boto3
botocore